Adding android R Compatible cc rules.
Currently pdfium have a dependency on libicu which is introduced in
android 31. Thus clients using pdfium running on android 30 or less
cannot include this. Adding the complile time flag which will hide api
that require android 31 when enabled. Thus ensuring clients running on
android 30 or less can also include this. The only drawback on running
pdfium on android 30 or less will be additional unicode support which is
supported by libicu library.

Test: Build and ran mts locally on android 30 and above.
Bug: 390327689
Change-Id: Iafb1abab34fbc69112b0efcb77ba234063657148